+++ src/Sulu/Bundle/AdminBundle/Controller/AdminController.php
@@ -64,6 +64,9 @@ class AdminController extends Controller
                     [
                         'name' => $this->container->getParameter('sulu_admin.name'),
                         'locales' => $this->container->getParameter('sulu_core.locales'),
+                        'translated_locales' => $this->container->getParameter('sulu_core.translated_locales'),
+                        'translations' => $this->container->getParameter('sulu_core.translations'),
+                        'fallback_locale' => $this->container->getParameter('sulu_core.fallback_locale'),
                         'suluVersion' => $this->container->getParameter('sulu.version'),
                         'user' => $user,
                         'config' => $jsconfig,