+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG for Sulu
 ==================
 
+* 1.0.6 (2015-08-05)
+    * HOTFIX      #1448 [AdminBundle]    Fixed additional system languages
+
 * 1.0.5 (2015-08-03)
     * HOTFIX      #--   [AdminBundle]    Fixed ckeditor overlay buttons for windows